Include Editors and Contributors
At the W3C Solid Community Group call on the 11th April it was decided to include the spec editor and contributors to each of the Solid specification repositories. The editor is clearly the Solid Specifications Repository Manager. 

What do you think should be the criteria to be included as a contributor?